+++ cli/csstools-cli/CHANGELOG.md
@@ -4,9 +4,11 @@
 
 - Emit warnings from plugins
 - Added `@csstools/postcss-is-pseudo-class`
+- Added `@csstools/postcss-hwb-function`
 
 ```sh
 npx -y @csstools/cli postcss-is-pseudo-class
+npx -y @csstools/cli postcss-hwb-function
 ```
 
 ### 1.0.0 (January 2, 2022)